refactor(tests): reorganize regression tests by module instead of severity
Move tests from severity-grouped files (test_high/medium/low_severity_fixes.py)
into module-grouped files under the existing test directory structure.

A maintainer looking for sarif_utils regression tests now finds them in
tests/unit/utils/test_sarif_utils_regression.py rather than hunting
across three severity files.

Cross-cutting invariant tests (mutable defaults, cwd defaults, environ
mutation, dead code removal) stay in tests/unit/ since they enforce
structural properties across the whole package.
