--- tests/unit/test_high_severity_fixes.py
@@ -1,383 +1,0 @@
-"""Tests for high-severity bug fixes.
-
-Each test is written RED first (to confirm the bug exists), then the
-corresponding source fix turns it GREEN.
-"""
-
-import os
-import re
-import shutil
-import sys
-import tempfile
-from pathlib import Path
-from unittest.mock import MagicMock, patch
-
-import pytest
-
-
-# ---------------------------------------------------------------------------
-# Bug #2 -- download_utils.py: code injection via f-string in
-#            create_url_download_command
-# ---------------------------------------------------------------------------
-class TestDownloadUtilsCodeInjection:
-    """A single quote in url/destination must not escape the Python string."""
-
-    def test_single_quote_in_url_no_injection(self):
-        """URL containing a single quote must not break the command args."""
-        from automated_security_helper.utils.download_utils import (
-            create_url_download_command,
-        )
-
-        evil_url = "https://example.com/bin'injection"
-        dest = tempfile.mkdtemp()
-        try:
-            cmd = create_url_download_command(url=evil_url, destination=dest)
-            # After the fix, the command should use --url/--dest flags or
-            # env vars instead of interpolating into a python -c string.
-            # No arg should contain the raw URL inside a python source string.
-            for arg in cmd.args:
-                if "-c" == arg:
-                    # If -c is still used, the url must not appear raw
-                    idx = cmd.args.index(arg)
-                    if idx + 1 < len(cmd.args):
-                        script = cmd.args[idx + 1]
-                        assert "'" + "injection" not in script or "sys.argv" in script or "os.environ" in script, (
-                            "URL interpolated raw into python -c script"
-                        )
-        finally:
-            shutil.rmtree(dest, ignore_errors=True)
-
-    def test_command_does_not_use_raw_interpolation(self):
-        """The command should pass values safely, not via f-string interpolation."""
-        from automated_security_helper.utils.download_utils import (
-            create_url_download_command,
-        )
-
-        url = "https://example.com/binary"
-        dest = tempfile.mkdtemp()
-        try:
-            cmd = create_url_download_command(url=url, destination=dest)
-            # After fix: url and dest should be passed as separate args, not
-            # interpolated into a python -c source string
-            if "-c" in cmd.args:
-                idx = cmd.args.index("-c")
-                script = cmd.args[idx + 1]
-                # The script should reference sys.argv, not contain the literal URL
-                assert url not in script, (
-                    f"URL appears literally in -c script: {script}"
-                )
-        finally:
-            shutil.rmtree(dest, ignore_errors=True)
-
-
-# ---------------------------------------------------------------------------
-# Bug #6 -- execution_engine.py: operator precedence on _max_workers
-# ---------------------------------------------------------------------------
-class TestMaxWorkersOperatorPrecedence:
-    """os.cpu_count() or 1 + 4 must parse as (os.cpu_count() or 1) + 4."""
-
-    def test_max_workers_with_cpu_count_available(self):
-        """When cpu_count() returns 8, _max_workers should be min(32, 12)."""
-        with patch("os.cpu_count", return_value=8):
-            result = min(32, (os.cpu_count() or 1) + 4)
-            assert result == 12  # 8 + 4
-
-    def test_max_workers_with_cpu_count_none(self):
-        """When cpu_count() returns None, _max_workers should be min(32, 5)."""
-        with patch("os.cpu_count", return_value=None):
-            result = min(32, (os.cpu_count() or 1) + 4)
-            assert result == 5  # (None or 1) + 4 = 5
-
-    def test_buggy_expression_demonstrates_problem(self):
-        """Demonstrate the bug: os.cpu_count() or 1 + 4 gives wrong result."""
-        with patch("os.cpu_count", return_value=8):
-            buggy = os.cpu_count() or 1 + 4  # parses as cpu_count() or 5
-            fixed = (os.cpu_count() or 1) + 4
-            assert buggy == 8  # bug: returns just cpu_count()
-            assert fixed == 12  # correct: cpu_count() + 4
-
-
-# ---------------------------------------------------------------------------
-# Bug #7 -- execution_engine.py: split delimiter mismatch ", " vs ","
-# ---------------------------------------------------------------------------
-class TestSplitDelimiterMismatch:
-    """Config values like 'a,b' (no space) must still split correctly."""
-
-    def test_split_comma_no_space(self):
-        """'a,b' split with ',' and strip gives ['a', 'b']."""
-        item = "a,b"
-        result = [subitem.strip() for subitem in item.split(",")]
-        assert result == ["a", "b"]
-
-    def test_split_comma_with_space(self):
-        """'a, b' split with ',' and strip gives ['a', 'b']."""
-        item = "a, b"
-        result = [subitem.strip() for subitem in item.split(",")]
-        assert result == ["a", "b"]
-
-    def test_split_comma_space_delimiter_fails_no_space(self):
-        """The OLD behavior: splitting 'a,b' on ', ' yields ['a,b'] -- unsplit."""
-        item = "a,b"
-        buggy_result = item.split(", ")
-        assert buggy_result == ["a,b"]  # confirms the bug
-
-
-# ---------------------------------------------------------------------------
-# Bug #10 -- orchestrator.py: shutil.rmtree(None) crash
-# ---------------------------------------------------------------------------
-class TestOrchestratorCleanupNoneWorkDir:
-    """Cleanup must not crash when work_dir is None."""
-
-    def test_cleanup_with_none_work_dir(self):
-        """shutil.rmtree(None) would raise TypeError; guard required."""
-        # Test the guard logic directly -- ASHScanOrchestrator is a Pydantic
-        # model that needs full init, so we verify the pattern in isolation.
-        work_dir = None
-        no_cleanup = False
-        if not no_cleanup:
-            if work_dir and Path(work_dir).exists():
-                shutil.rmtree(work_dir)
-        # If we got here without TypeError, the guard works.
-
-    def test_original_code_would_crash(self):
-        """Demonstrate that shutil.rmtree(None) raises TypeError."""
-        with pytest.raises(TypeError):
-            shutil.rmtree(None)
-
-    def test_cleanup_with_nonexistent_work_dir(self):
-        """A path that doesn't exist should also be guarded."""
-        work_dir = Path("/tmp/nonexistent_ash_work_dir_xyz_test")
-        if work_dir.exists():
-            shutil.rmtree(work_dir)
-        # Guard should prevent rmtree on nonexistent path
-        if work_dir and Path(work_dir).exists():
-            shutil.rmtree(work_dir)
-
-
-# ---------------------------------------------------------------------------
-# Bug #15 -- scan_tracking.py: severity key case mismatch
-# ---------------------------------------------------------------------------
-class TestSeverityKeyCaseMismatch:
-    """extract_findings_summary must handle uppercase severity values."""
-
-    def test_uppercase_severity_counted(self):
-        """Findings with uppercase 'CRITICAL' must be counted."""
-        from automated_security_helper.core.resource_management.scan_tracking import (
-            extract_findings_summary,
-        )
-
-        findings = [
-            {"severity": "CRITICAL"},
-            {"severity": "CRITICAL"},
-            {"severity": "HIGH"},
-            {"severity": "MEDIUM"},
-            {"severity": "LOW"},
-            {"severity": "INFO"},
-        ]
-        summary = extract_findings_summary(findings)
-        assert summary["critical"] == 2
-        assert summary["high"] == 1
-        assert summary["medium"] == 1
-        assert summary["low"] == 1
-        assert summary["info"] == 1
-
-    def test_lowercase_severity_counted(self):
-        """Findings with lowercase severity must also work."""
-        from automated_security_helper.core.resource_management.scan_tracking import (
-            extract_findings_summary,
-        )
-
-        findings = [
-            {"severity": "critical"},
-            {"severity": "high"},
-        ]
-        summary = extract_findings_summary(findings)
-        assert summary["critical"] == 1
-        assert summary["high"] == 1
-
-    def test_mixed_case_severity(self):
-        """Mixed case like 'Critical' must be counted."""
-        from automated_security_helper.core.resource_management.scan_tracking import (
-            extract_findings_summary,
-        )
-
-        findings = [
-            {"severity": "Critical"},
-            {"severity": "HIGH"},
-            {"severity": "medium"},
-        ]
-        summary = extract_findings_summary(findings)
-        assert summary["critical"] == 1
-        assert summary["high"] == 1
-        assert summary["medium"] == 1
-
-
-# ---------------------------------------------------------------------------
-# Bug #19 -- get_reporter_mappings.py: __class__.__name__ on class object
-# ---------------------------------------------------------------------------
-class TestReporterClassName:
-    """reporter.__class__.__name__ on a class yields 'type', not the class name."""
-
-    def test_class_name_not_type(self):
-        """Calling __class__.__name__ on a class returns 'type'."""
-
-        class FakeReporter:
-            pass
-
-        # Bug behavior
-        assert FakeReporter.__class__.__name__ == "type"
-        # Correct behavior
-        assert FakeReporter.__name__ == "FakeReporter"
-
-    def test_reporter_mapping_uses_correct_name(self):
-        """After fix, get_reporter_mappings should use the actual class name."""
-
-        class SarifReporter:
-            pass
-
-        # Wrong: class object's __class__ is the metaclass 'type'
-        wrong_name = SarifReporter.__class__.__name__
-        assert wrong_name == "type"
-
-        # Right: the class's own __name__
-        correct_name = SarifReporter.__name__
-        assert correct_name == "SarifReporter"
-
-
-# ---------------------------------------------------------------------------
-# Bug #22 -- sarif_utils.py: path_matches_pattern substring check backwards
-# ---------------------------------------------------------------------------
-class TestPathMatchesPattern:
-    """'path in pat' should be 'pat in path' (or fnmatch)."""
-
-    def test_pattern_matches_subpath(self):
-        """Pattern 'src/foo' should match path 'src/foo/bar.py'."""
-        from automated_security_helper.utils.sarif_utils import path_matches_pattern
-
-        assert path_matches_pattern("src/foo/bar.py", "src/foo") is True
-
-    def test_path_does_not_match_longer_pattern(self):
-        """Path 'foo' must NOT match pattern 'src/foo/bar.py'."""
-        from automated_security_helper.utils.sarif_utils import path_matches_pattern
-
-        # With the bug: 'foo' in 'src/foo/bar.py' would be True
-        assert path_matches_pattern("foo", "src/foo/bar.py") is False
-
-    def test_exact_match_still_works(self):
-        """Exact match should still return True."""
-        from automated_security_helper.utils.sarif_utils import path_matches_pattern
-
-        assert path_matches_pattern("src/foo.py", "src/foo.py") is True
-
-
-# ---------------------------------------------------------------------------
-# Bug #24 -- version_management.py: re.sub replaces all occurrences
-# ---------------------------------------------------------------------------
-class TestVersionReSubCount:
-    """re.sub without count=1 replaces all 'version = ...' lines."""
-
-    def test_only_first_version_replaced(self):
-        """Only the project version should be replaced, not dependency pins."""
-        content = (
-            '[project]\nname = "ash"\nversion = "1.0.0"\n\n'
-            '[dependencies]\nfoo = {version = "2.0.0"}\n'
-            'bar = {version = "3.0.0"}\n'
-        )
-        pattern = r'(version\s*=\s*")([^"]+)(")'
-        # Simulate the fix: count=1
-        new_content = re.sub(pattern, r'\g<1>9.9.9\g<3>', content, count=1)
-        # First version replaced
-        assert 'version = "9.9.9"' in new_content
-        # Dependency versions preserved
-        assert 'version = "2.0.0"' in new_content
-        assert 'version = "3.0.0"' in new_content
-
-    def test_buggy_replaces_all(self):
-        """Without count=1, all version lines get replaced."""
-        content = (
-            '[project]\nversion = "1.0.0"\n'
-            '[deps]\nfoo = {version = "2.0.0"}\n'
-        )
-        pattern = r'(version\s*=\s*")([^"]+)(")'
-        buggy = re.sub(pattern, r'\g<1>9.9.9\g<3>', content)
-        # Bug: both are replaced
-        assert buggy.count('version = "9.9.9"') == 2
-
-
-# ---------------------------------------------------------------------------
-# Bug #156/157/158 -- secret_masking.py
-# ---------------------------------------------------------------------------
-class TestSecretMasking:
-    """Tests for secret masking deficiencies."""
-
-    # Bug #156: 4-char secrets reveal 50% (show first + last of 4)
-    def test_short_secret_masking_ratio(self):
-        """4-char secret must have >= 75% masked characters."""
-        from automated_security_helper.utils.secret_masking import _mask_secret_value
-
-        masked = _mask_secret_value("abcd")
-        star_count = masked.count("*")
-        total = len(masked)
-        ratio = star_count / total
-        assert ratio >= 0.75, f"Only {ratio:.0%} masked for 4-char secret: {masked}"
-
-    def test_five_char_secret_masking(self):
-        """5-char secret must have >= 75% masked."""
-        from automated_security_helper.utils.secret_masking import _mask_secret_value
-
-        masked = _mask_secret_value("abcde")
-        star_count = masked.count("*")
-        ratio = star_count / len(masked)
-        assert ratio >= 0.75, f"Only {ratio:.0%} masked for 5-char secret: {masked}"
-
-    def test_ten_char_secret_masking(self):
-        """10-char secret masking should be reasonable."""
-        from automated_security_helper.utils.secret_masking import _mask_secret_value
-
-        masked = _mask_secret_value("abcdefghij")
-        star_count = masked.count("*")
-        ratio = star_count / len(masked)
-        assert ratio >= 0.50, f"Only {ratio:.0%} masked for 10-char secret"
-
-    # Bug #158: B106 and B107 are not dispatched
-    def test_b106_rule_masks_secret(self):
-        """B106 (hardcoded password in funcarg) must be masked."""
-        from automated_security_helper.utils.secret_masking import mask_secret_in_text
-
-        text = "Possible hardcoded password: 'SuperSecret123'"
-        result = mask_secret_in_text(text, rule_id="B106")
-        assert "SuperSecret123" not in result
-
-    def test_b107_rule_masks_secret(self):
-        """B107 (hardcoded password in config) must be masked."""
-        from automated_security_helper.utils.secret_masking import mask_secret_in_text
-
-        text = "Possible hardcoded password: 'MyPassw0rd'"
-        result = mask_secret_in_text(text, rule_id="B107")
-        assert "MyPassw0rd" not in result
-
-    def test_b105_still_works(self):
-        """B105 masking should continue to work after the fix."""
-        from automated_security_helper.utils.secret_masking import mask_secret_in_text
-
-        text = "Possible hardcoded password: 'secret123'"
-        result = mask_secret_in_text(text, rule_id="B105")
-        assert "secret123" not in result
-
-    # Bug #157: regex stops at internal quote
-    def test_secret_with_internal_quote(self):
-        """Secret containing a quote should still be fully masked."""
-        from automated_security_helper.utils.secret_masking import mask_secret_in_text
-
-        text = "Possible hardcoded password: 'it\\'s_a_secret'"
-        result = mask_secret_in_text(text, rule_id="B105")
-        # The secret value should not appear in cleartext
-        assert "it's_a_secret" not in result or "****" in result
-
-    def test_secret_with_double_quote_delimiter(self):
-        """Secret in double quotes should be masked."""
-        from automated_security_helper.utils.secret_masking import mask_secret_in_text
-
-        text = 'Possible hardcoded password: "MySecret99"'
-        result = mask_secret_in_text(text, rule_id="B105")
-        assert "MySecret99" not in result
